+++ types/index.d.ts
@@ -347,7 +347,7 @@ declare class RefreshToken extends BaseToken {
     sessionUid?: string;
     expiresWithSession?: boolean;
     'x5t#S256'?: string;
-    'jkt#S256'?: string;
+    jkt?: string;
     grantId: string;
     gty: string;
     [key: string]: any;
@@ -367,7 +367,7 @@ declare class RefreshToken extends BaseToken {
   sessionUid?: string;
   expiresWithSession?: boolean;
   'x5t#S256'?: string;
-  'jkt#S256'?: string;
+  jkt?: string;
   grantId?: string;
   gty?: string;
   consumed: any;
@@ -395,7 +395,7 @@ declare class AuthorizationCode extends BaseToken {
     sessionUid?: string;
     expiresWithSession?: boolean;
     'x5t#S256'?: string;
-    'jkt#S256'?: string;
+    jkt?: string;
     grantId: string;
     gty: string;
     [key: string]: any;
@@ -416,7 +416,7 @@ declare class AuthorizationCode extends BaseToken {
   sessionUid?: string;
   expiresWithSession?: boolean;
   'x5t#S256'?: string;
-  'jkt#S256'?: string;
+  jkt?: string;
   grantId?: string;
   gty?: string;
 
@@ -432,6 +432,9 @@ declare class DeviceCode extends BaseToken {
     deviceInfo: AnyObject;
     [key: string]: any;
   });
+
+  static findByUserCode(userCode: string, options?: { ignoreExpiration?: boolean }): Promise<DeviceCode | undefined>;
+
   readonly kind: 'DeviceCode';
   error?: string;
   errorDescription?: string;
@@ -472,7 +475,7 @@ declare class ClientCredentials extends BaseToken {
   aud: string | string[];
   readonly tokenType: string;
   'x5t#S256'?: string;
-  'jkt#S256'?: string;
+  jkt?: string;
 
   setAudiences(audience: string | string[]): void;
   isSenderConstrained(): boolean;
@@ -505,7 +508,7 @@ declare class AccessToken extends BaseToken {
     sessionUid?: string;
     expiresWithSession?: boolean;
     'x5t#S256'?: string;
-    'jkt#S256'?: string;
+    jkt?: string;
     grantId: string;
     gty: string;
     [key: string]: any;
@@ -523,7 +526,7 @@ declare class AccessToken extends BaseToken {
   expiresWithSession?: boolean;
   readonly tokenType: string;
   'x5t#S256'?: string;
-  'jkt#S256'?: string;
+  jkt?: string;
 
   setAudiences(audience: string | string[]): void;
   isSenderConstrained(): boolean;
@@ -777,7 +780,7 @@ export interface AdapterPayload {
   transient?: boolean;
   uid?: string;
   userCode?: string;
-  'jkt#S256'?: string;
+  jkt?: string;
   'jwt-ietf'?: string;
   'x5t#S256'?: string;
 }
@@ -893,7 +896,7 @@ export interface Configuration {
         whitelist?: string[] | Set<string>;
       };
     };
-    dPoP?: { enabled?: boolean, iatTolerance?: number, ack?: 'id-02' },
+    dPoP?: { enabled?: boolean, iatTolerance?: number, ack?: 'id-03' },
 
     sessionManagement?: { enabled?: boolean, keepHeaders?: boolean, ack?: 28 },
 
@@ -1093,6 +1096,8 @@ export class Provider extends events.EventEmitter {
   env?: Koa['env'];
   proxy?: Koa['proxy'];
   subdomainOffset?: number;
+  proxyIpHeader?: string;
+  maxIpsCount?: number;
   keys?: Koa['keys'];
   listen: Koa['listen'];